pyproject.toml
[tool.poetry]
name = "housing-utilities"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
celery = "^5.4.0"
redis = "^5.1.1"
django-celery-beat = "^2.7.0"
eventlet = "^0.37.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
